\name{AIC.softmax}
\alias{AIC.softmax}
\alias{AIC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate AIC of Fitted Softmax Regression Model
}
\description{
Caculate akaike information criterion of fitted softmax regression model
}
\usage{
\method{AIC}{softmax}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A object of "softmax" returned by \code{\link{softmaxReg}} function
}
\item{\dots}{Other arguments}
}
\value{
Numeric Value of AIC
}
\seealso{
\code{\link{BIC.softmax}}
\code{\link{softmaxReg}}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
