\name{BIC.softmax}
\alias{BIC.softmax}
\alias{BIC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate BIC of Fitted Softmax Regression Model
}
\description{
Caculate bayesian information criterion of fitted softmax regression model
}
\usage{
\method{BIC}{softmax}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A object of "softmax" returned by \code{\link{softmaxReg}} function
}
\item{\dots}{Other arguments}
}
\value{
Numeric Value of BIC
}
\seealso{
\code{\link{AIC.softmax}}
\code{\link{softmaxReg}}
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
