\name{softmaxreg-internal}
\alias{activate}
\alias{activateDeri}
\alias{ReLuDeri}
\alias{ReLuFunc}
\alias{sigmoid}
\alias{sigmoidDeri}
\alias{softmax}
\alias{softmaxFunc}
\alias{softmaxLoss}
\alias{softmaxProp}
\alias{tanhFunc}
\alias{tanhDeri}
\alias{MSELoss}
\alias{addL2GradPenalty}
\alias{addL2LossPenalty}
\alias{getRandomBatch}
\alias{calcGrad}
\alias{calcLoss}
\alias{checkGradMode}
\alias{convertList2Matrix}
\alias{copyShape}
\alias{forwardProp}
\alias{forwardPropParallel}
\alias{getLogValue}
\alias{gradientDescent}
\alias{updatePara}
\alias{initPara}
\alias{addList}
\alias{subtractList}
\alias{divideList}
\alias{dividePara}
\alias{divideListByList}
\alias{multiplyList}
\alias{multiplyListByList}
\alias{rootList}
\alias{squareList}
\alias{isInteger}
\alias{calcInfoCriteria}
\alias{load_image_file}
\alias{load_label_file}
\alias{show_digit}
\title{Internal Softmaxreg Functions}
\description{
 Internal Softmaxreg functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
